Updated parallelnew code block directive
Allows the user to hover #pragma omp parallelnew over a code block of their choice to produce nondeterministic output of data structures.  This is achieved through a recursive lambda function that searches the MCXX AST for said structure, and alters it via the user's choosing.  Here we demonstrate a test integer and double its value.  OpenMP functionality is preserved just as one would observe in #pragma omp parallel.